Replace ints with strong type aliases for registers and labels
This commit is a major refactor that hides the int representation
of registers and labels. It creates a new module called Ir that
holds common types used by multiple "backend" modules. In the
process of refactoring, I caught a bug in which I used a register
as a color, so strong typing and information hiding does pay off.
:)

This commit also moves the definition of the operand type to
ir.ml.