Give Anf module its own distinct file/package type
This commit gives the ANF intermediate representation its own type
for a "toplevel" file or package. The purpose of this commit was
to store the toplevel register generator somewhere after
generating ANF. However, the process of making this change led to
a larger refactor that reorganized the compiler pipeline to better
separate the different phases. Now, we don't need to keep each
phase's IR around in order to generate the ANF.