Support the importing of types from other files
This commit adds the feature of importing types from other files. The
changes in this commit involves the transfer of code from the typecheck
module to the desugar module. The transferred code deals with the
translation of types from the ast.ml representation to the type.ml
representation and the processing of algebraic data type definitions.