Make explicit type for polytypes
Before this commit, polytypes / type schemes and monotypes were represented
by the same type, Type.t. A polytype was merely a type that contained rigid
type variables. This commit adds a separate type for polytypes. This type
lists the type variables that the polytype quantifies over.

When instantiating a polytype, only the type variables that are explicitly
quantified over are replaced with fresh wobbly types.

In addition, this commit makes the typechecker generalization function make
a deep copy of the type body instead of having multiple polytypes share the
same body. This change is necessary to make instantiation work when multiple
types being generalized share the same type variables.